% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phild.R
\name{phild}
\alias{phild}
\title{phild}
\usage{
phild(
  lambda = NULL,
  alpha = NULL,
  beta = NULL,
  rho = NULL,
  betax = NULL,
  distribution = "weibull",
  nom_loglik = "logcomplet",
  dataset = NULL,
  NomCovariable = NULL
)
}
\arguments{
\item{lambda}{Vector of initial values for the scale parameters of the
baseline intensity functions for the considered distributions (Weibull,
Gompertz, and generalized log-logistic).}

\item{alpha}{Vector of initial values for the shape parameters of the
considered distributions (Weibull, Gompertz, and generalized log-logistic).}

\item{beta}{Initial value of the proportionality parameter.
This argument is used only for estimating the constrained
illness--death model.}

\item{rho}{Vector of initial values for the location parameters of the
generalized log-logistic distribution.}

\item{betax}{Matrix of initial values for covariate coefficients.}

\item{distribution}{Name of the model used. Possible values are:
Weibull (\code{"weibull"}), Gompertz (\code{"gompertz"}), and
generalized log-logistic (\code{"gll"}). The default model is Weibull.}

\item{nom_loglik}{Specifies the illness--death model to be estimated:
\code{"logcomplet"} for the unconstrained model and
\code{"logincomplet"} for the constrained model.
The default is the unconstrained model.}

\item{dataset}{Dataset used for estimation.}

\item{NomCovariable}{Vector containing the names of the covariates
included in the dataset.}
}
\value{
The function returns the maximum likelihood estimators of the
transition intensity functions, as well as their variances and
95\% confidence intervals.
}
\description{
Estimates the parameters of constrained and unconstrained illness--death models
using the maximum likelihood method, considering Weibull, Gompertz, and
generalized log-logistic distributions.
}
\examples{
library(survival)
library(eha)

data(heart)

dataset <- transILD(
  heart,
  identifiant = "id",
  observation = "event",
  indicateur = "transplant",
  tempsUn = "start",
  tempsDeux = "stop",
  format_long = FALSE
)

M <- phild(
  lambda = rep(0.001, 3),
  alpha = rep(0.001, 3),
  betax = matrix(
    c(rep(-0.001, 3),
      rep(-0.001, 3),
      rep(-0.001, 3)),
    3, 3,
    byrow = TRUE
  ),
  distribution = "weibull",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = c("age", "year", "surgery")
)

M

phreg(
  Surv(start, stop, status) ~ age + year + surgery,
  data = dataset[(dataset$from == 1 & dataset$to == 3), ],
  dist = "weibull"
)

M$`Transition 13`

}
\seealso{
phildII
}
