% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phildII.R
\name{phildII}
\alias{phildII}
\title{phildII}
\usage{
phildII(
  Lambda = NULL,
  betax = NULL,
  beta = NULL,
  distribution = "pwexp",
  nom_loglik = "logcomplet",
  dataset = NULL,
  NomCovariable = NULL,
  breakpoints = NULL
)
}
\arguments{
\item{Lambda}{Matrice des valeurs initiales des paramètres de dimension
$(2 * 3)$.}

\item{betax}{Matrice des valeurs initiales des coefficients des covariables.}

\item{beta}{Valeur initiale du paramètre de proportionnalité.
Cet argument est utilisé uniquement pour l'estimation du modèle
maladie--décès contraint.}

\item{distribution}{Nom du modèle utilisé. Les valeurs possibles sont :
exponentielle constante par morceaux (\code{"pwexp"}) et
approximation linéaire par morceaux (\code{"pla"}).
Le modèle par défaut est \code{"pwexp"}.}

\item{nom_loglik}{Indique le modèle maladie--décès à estimer :
\code{"logcomplet"} pour le modèle non contraint et
\code{"logincomplet"} pour le modèle contraint.
Le modèle par défaut est le modèle non contraint.}

\item{dataset}{Base de données utilisée pour l'estimation.}

\item{NomCovariable}{Vecteur contenant les noms des covariables
présentes dans la base de données.}

\item{breakpoints}{Vecteur des points de coupure pour l'approximation
linéaire par morceaux ou matrice des intervalles de coupure pour
l'exponentielle constante par morceaux.}
}
\value{
La fonction retourne les estimateurs du maximum de vraisemblance
des fonctions d'intensité de transition, ainsi que leurs variances et
leurs intervalles de confiance au niveau de $95\%$.
}
\description{
Estime les paramètres des modèles maladie--décès contraint et non contraint
par la méthode du maximum de vraisemblance en considérant l'exponentielle
constante par morceaux et l'approximation linéaire par morceaux.
}
\examples{
library(PWEXP)

dataset <- simILD(
  n = 1000,
  censure = 0.2,
  lambda = t(cbind(c(0.18, 0.2),
                    c(0.15, 0.2),
                    c(0.15, 0.2))),
  breakpoints = 3,
  beta = matrix(c(0.4, 0.2, 0.2), 1, 3, byrow = TRUE),
  covariable = 1,
  distribution = "pwexp",
  lat_long = TRUE
)

phildII(
  Lambda = cbind(rep(0.001, 2),
                  rep(0.001, 2),
                  rep(0.001, 2)),
  betax = matrix(rep(-0.001, 3), 1, 3, byrow = TRUE),
  distribution = "pwexp",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = "X",
  breakpoints = cbind(c(0, 3), c(3, Inf))
)

phildII(
  Lambda = cbind(rep(0.001, 2),
                  rep(0.001, 2),
                  rep(0.001, 2)),
  betax = matrix(rep(-0.001, 3), 1, 3, byrow = TRUE),
  distribution = "pla",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = "X",
  breakpoints = c(0, 3, Inf)
)

}
\seealso{
phild
}
