% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phildtestII.R
\name{phildtestII}
\alias{phildtestII}
\title{phildtestII}
\usage{
phildtestII(
  Lambda,
  beta = NULL,
  betax = NULL,
  distribution = "pwexp",
  dataset = NULL,
  breakpoints,
  NomCovariable = NULL
)
}
\arguments{
\item{Lambda}{Matrice des valeurs initiales des paramètres de dimension
$(2 * 3)$.}

\item{beta}{Valeur initiale du paramètre de proportionnalité.
Cet argument est utilisé uniquement pour l'estimation du modèle
maladie--décès contraint.}

\item{betax}{Matrice des valeurs initiales des coefficients des covariables.}

\item{distribution}{Nom du modèle utilisé. Les valeurs possibles sont :
exponentielle constante par morceaux (\code{"pwexp"}) et
approximation linéaire par morceaux (\code{"pla"}).
Le modèle par défaut est \code{"pwexp"}.}

\item{dataset}{Base de données utilisée pour l'estimation.}

\item{breakpoints}{Vecteur des points de coupure pour l'approximation
linéaire par morceaux ou matrice des intervalles pour
l'exponentielle constante par morceaux.}

\item{NomCovariable}{Vecteur contenant les noms des covariables
présentes dans la base de données.}
}
\value{
La fonction retourne la statistique du test du rapport de vraisemblance
ainsi que la p-value associée.
}
\description{
Effectue le test du rapport de vraisemblance entre le modèle maladie--décès
contraint et le modèle non contraint, en considérant l'exponentielle
constante par morceaux et l'approximation linéaire par morceaux.
}
\seealso{
phildtest
}
