% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transILD.R
\name{transILD}
\alias{transILD}
\title{transILD}
\usage{
transILD(
  initial_data,
  identifiant = NULL,
  observation = NULL,
  indicateur = NULL,
  tempsUn = NULL,
  tempsDeux = NULL,
  format_long
)
}
\arguments{
\item{initial_data}{Base de données sous format court ou long.}

\item{identifiant}{Nom de la colonne correspondant à l'identifiant
de chaque individu.}

\item{observation}{Nom de la colonne correspondant à l'indicateur
d'événement.}

\item{indicateur}{Nom de la colonne correspondant à l'indicateur
de maladie.}

\item{tempsUn}{Nom de la colonne correspondant au temps d'entrée
dans un état $k$.}

\item{tempsDeux}{Nom de la colonne correspondant au temps de sortie
d'un état $k$.}

\item{format_long}{Indique si la base de données est déjà sous
format long (\code{TRUE}) ou sous format court (\code{FALSE}).}
}
\value{
La fonction retourne une base de données contenant les variables :
\code{id} (identifiant de l'individu), \code{start} (temps d'entrée dans l'état $k$),
\code{stop} (temps de sortie de l'état $k$), \code{status} (indique si la transition
$k -> l$ est observée), \code{from} (état de départ), \code{to} (état d'arrivée),
ainsi que les covariables présentes dans la base de données initiale.
}
\description{
Transforme une base de données sous un format long compatible avec
les fonctions du package. Les fonctions de vraisemblance du package
sont construites en supposant que les données sont organisées sous
un format long. Cette fonction permet également de reformater toute base de données,
qu'elle soit initialement sous format court ou long, selon la structure
requise par les fonctions du package. Nous recommandons donc aux utilisateurs
d'utiliser préalablement la fonction \code{transILD} avant de passer un jeu
de données aux fonctions \code{phild}, \code{phildII},
\code{phildtest} et \code{phildtestII}.
}
\examples{
library(survival)
data(heart)

transILD(
  heart,
  identifiant = "id",
  observation = "event",
  indicateur = "transplant",
  tempsUn = "start",
  tempsDeux = "stop",
  format_long = FALSE
)

}
