% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phild.R
\name{phild}
\alias{phild}
\title{phild}
\usage{
phild(
  lambda = NULL,
  alpha = NULL,
  beta = NULL,
  rho = NULL,
  betax = NULL,
  distribution = "weibull",
  nom_loglik = "logcomplet",
  dataset = NULL,
  NomCovariable = NULL
)
}
\arguments{
\item{lambda}{Vecteur des valeurs initiales des paramètres d'échelle des
fonctions d'intensité de base pour les lois considérées (Weibull, Gompertz
et log-logistique généralisée).}

\item{alpha}{Vecteur des valeurs initiales des paramètres de forme
pour les lois considérées (Weibull, Gompertz et log-logistique généralisée).}

\item{beta}{Valeur initiale du paramètre de proportionnalité.
Cet argument est utilisé uniquement pour l'estimation du modèle
maladie--décès contraint.}

\item{rho}{Vecteur des valeurs initiales des paramètres de position
de la loi log-logistique généralisée.}

\item{betax}{Matrice des valeurs initiales des coefficients des covariables.}

\item{distribution}{Nom du modèle utilisé. Les valeurs possibles sont :
Weibull (\code{"weibull"}), Gompertz (\code{"gompertz"}) et
log-logistique généralisée (\code{"gll"}). Le modèle par défaut est Weibull.}

\item{nom_loglik}{Indique le modèle maladie--décès à estimer :
\code{"logcomplet"} pour le modèle non contraint et
\code{"logincomplet"} pour le modèle contraint.
Le modèle par défaut est le modèle non contraint.}

\item{dataset}{Base de données utilisée pour l'estimation.}

\item{NomCovariable}{Vecteur contenant les noms des covariables
présentes dans la base de données.}
}
\value{
La fonction retourne les estimateurs du maximum de vraisemblance
des fonctions d'intensité de transition, ainsi que leurs variances et
leurs intervalles de confiance au niveau de $95\%$.
}
\description{
Estime les paramètres des modèles maladie--décès contraint et non contraint
par la méthode du maximum de vraisemblance, en considérant les lois de Weibull,
de Gompertz et log-logistique généralisée.
}
\examples{
library(survival)
library(eha)

data(heart)

dataset <- transILD(
  heart,
  identifiant = "id",
  observation = "event",
  indicateur = "transplant",
  tempsUn = "start",
  tempsDeux = "stop",
  format_long = FALSE
)

M <- phild(
  lambda = rep(0.001, 3),
  alpha = rep(0.001, 3),
  betax = matrix(
    c(rep(-0.001, 3),
      rep(-0.001, 3),
      rep(-0.001, 3)),
    3, 3,
    byrow = TRUE
  ),
  distribution = "weibull",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = c("age", "year", "surgery")
)

M

phreg(
  Surv(start, stop, status) ~ age + year + surgery,
  data = dataset[(dataset$from == 1 & dataset$to == 3), ],
  dist = "weibull"
)

M$`Transition 13`

}
\seealso{
phildII
}
