% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phildII.R
\name{phildII}
\alias{phildII}
\title{phildII}
\usage{
phildII(
  Lambda = NULL,
  betax = NULL,
  beta = NULL,
  distribution = "pwexp",
  nom_loglik = "logcomplet",
  dataset = NULL,
  NomCovariable = NULL,
  breakpoints = NULL
)
}
\arguments{
\item{Lambda}{Matrix of initial parameter values of dimension (2 * 3).}

\item{betax}{Matrix of initial values for covariate coefficients.}

\item{beta}{Initial value of the proportionality parameter.
This argument is used only for estimating the constrained
illness--death model.}

\item{distribution}{Name of the model used. Possible values are:
piecewise constant exponential (\code{"pwexp"}) and
piecewise linear approximation (\code{"pla"}).
The default model is \code{"pwexp"}.}

\item{nom_loglik}{Specifies the illness--death model to be estimated:
\code{"logcomplet"} for the unconstrained model and
\code{"logincomplet"} for the constrained model.
The default is the unconstrained model.}

\item{dataset}{Dataset used for estimation.}

\item{NomCovariable}{Vector containing the names of the covariates
included in the dataset.}

\item{breakpoints}{Vector of cut points for the piecewise linear approximation
or matrix of cut intervals for the piecewise constant exponential model.}
}
\value{
The function returns the maximum likelihood estimators of the
transition intensity functions, as well as their variances and
95\% confidence intervals.
}
\description{
Estimates the parameters of constrained and unconstrained illness--death models
using the maximum likelihood method, considering the piecewise constant exponential
model and the piecewise linear approximation.
}
\examples{
library(PWEXP)

dataset <- simILD(
  n = 1000,
  censure = 0.2,
  lambda = t(cbind(c(0.18, 0.2),
                    c(0.15, 0.2),
                    c(0.15, 0.2))),
  breakpoints = 3,
  beta = matrix(c(0.4, 0.2, 0.2), 1, 3, byrow = TRUE),
  covariable = 1,
  distribution = "pwexp",
  lat_long = TRUE
)

phildII(
  Lambda = cbind(rep(0.001, 2),
                  rep(0.001, 2),
                  rep(0.001, 2)),
  betax = matrix(rep(-0.001, 3), 1, 3, byrow = TRUE),
  distribution = "pwexp",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = "X",
  breakpoints = cbind(c(0, 3), c(3, Inf))
)

phildII(
  Lambda = cbind(rep(0.001, 2),
                  rep(0.001, 2),
                  rep(0.001, 2)),
  betax = matrix(rep(-0.001, 3), 1, 3, byrow = TRUE),
  distribution = "pla",
  nom_loglik = "logcomplet",
  dataset = dataset,
  NomCovariable = "X",
  breakpoints = c(0, 3, Inf)
)

}
\seealso{
phild
}
