% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phildtest.R
\name{phildtest}
\alias{phildtest}
\title{phildtest}
\usage{
phildtest(
  lambda = NULL,
  alpha = NULL,
  rho = NULL,
  beta = NULL,
  betax = NULL,
  distribution = "weibull",
  dataset = NULL,
  NomCovariable = NULL
)
}
\arguments{
\item{lambda}{Vector of initial values for the scale parameters of the
baseline intensity functions for the considered distributions (Weibull,
Gompertz, and generalized log-logistic).}

\item{alpha}{Vector of initial values for the shape parameters of the
considered distributions (Weibull, Gompertz, and generalized log-logistic).}

\item{rho}{Vector of initial values for the location parameters of the
generalized log-logistic distribution.}

\item{beta}{Initial value of the proportionality parameter.
This argument is used only for estimating the constrained
illness--death model.}

\item{betax}{Matrix of initial values for covariate coefficients.}

\item{distribution}{Name of the model used. Possible values are:
Weibull (\code{"weibull"}), Gompertz (\code{"gompertz"}), and
generalized log-logistic (\code{"gll"}). The default model is Weibull.}

\item{dataset}{Dataset used for estimation.}

\item{NomCovariable}{Vector containing the names of the covariates
included in the dataset.}
}
\value{
The function returns the likelihood ratio test statistic
along with the associated p-value.
}
\description{
Performs a likelihood ratio test between the constrained and unconstrained
illness--death models, considering Weibull, Gompertz, and generalized
log-logistic distributions.
}
\seealso{
phildII
}
