% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phildtestII.R
\name{phildtestII}
\alias{phildtestII}
\title{phildtestII}
\usage{
phildtestII(
  Lambda,
  beta = NULL,
  betax = NULL,
  distribution = "pwexp",
  dataset = NULL,
  breakpoints,
  NomCovariable = NULL
)
}
\arguments{
\item{Lambda}{Matrix of initial parameter values of dimension (2 * 3).}

\item{beta}{Initial value of the proportionality parameter.
This argument is used only for estimating the constrained
illness--death model.}

\item{betax}{Matrix of initial values for covariate coefficients.}

\item{distribution}{Name of the model used. Possible values are:
piecewise constant exponential (\code{"pwexp"}) and
piecewise linear approximation (\code{"pla"}).
The default model is \code{"pwexp"}.}

\item{dataset}{Dataset used for estimation.}

\item{breakpoints}{Vector of cut points for the piecewise linear approximation
or matrix of intervals for the piecewise constant exponential model.}

\item{NomCovariable}{Vector containing the names of the covariates
included in the dataset.}
}
\value{
The function returns the likelihood ratio test statistic
along with the associated p-value.
}
\description{
Performs a likelihood ratio test between the constrained and unconstrained
illness--death models, considering the piecewise constant exponential model
and the piecewise linear approximation.
}
\seealso{
phildtest
}
