% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transILD.R
\name{transILD}
\alias{transILD}
\title{transILD}
\usage{
transILD(
  initial_data,
  identifiant = NULL,
  observation = NULL,
  indicateur = NULL,
  tempsUn = NULL,
  tempsDeux = NULL,
  format_long
)
}
\arguments{
\item{initial_data}{Dataset in short or long format.}

\item{identifiant}{Name of the column corresponding to the individual
identifier.}

\item{observation}{Name of the column corresponding to the event
indicator.}

\item{indicateur}{Name of the column corresponding to the illness
indicator.}

\item{tempsUn}{Name of the column corresponding to the entry time
into state k.}

\item{tempsDeux}{Name of the column corresponding to the exit time
from state k.}

\item{format_long}{Indicates whether the dataset is already in
long format (\code{TRUE}) or in short format (\code{FALSE}).}
}
\value{
The function returns a dataset containing the variables:
\code{id} (individual identifier), \code{start} (entry time into state k),
\code{stop} (exit time from state k), \code{status} (indicates whether the
transition k -> l is observed), \code{from} (starting state), \code{to}
(destination state), as well as the covariates present in the initial dataset.
}
\description{
Transforms a dataset into a long format compatible with the package functions.
The likelihood functions implemented in the package are constructed under the
assumption that the data are organized in long format. This function also allows
any dataset, whether initially in short or long format, to be reformatted
according to the structure required by the package functions. We therefore
recommend that users first apply the \code{transILD} function before passing
a dataset to the functions \code{phild}, \code{phildII},
\code{phildtest}, and \code{phildtestII}.
}
\examples{
library(survival)
data(heart)

transILD(
  heart,
  identifiant = "id",
  observation = "event",
  indicateur = "transplant",
  tempsUn = "start",
  tempsDeux = "stop",
  format_long = FALSE
)

}
